Store site-specific settings.
Site options can be stored for all sites, only for sites that use non-default
options, or never (the default is to save for non-defaults). A simple hash
table is used to store the data, because Chrome sync storage is limited to 2048
bytes per key and 512 keys. 419 keys are used for the hash table (and one for
the global settings), leaving 92 keys for future use.